+++ src/ads/ad.js
@@ -7,6 +7,7 @@
  */
 class Ad {
   _id: string;
+  _system: ?string;
   _url: ?string;
   _contentType: ?string;
   _title: ?string;
@@ -15,13 +16,15 @@ class Ad {
   _clickThroughUrl: ?string;
   _posterUrl: ?string;
   _skipOffset: ?number;
-  _linear: ?boolean;
+  _linear: boolean;
   _width: number;
   _height: number;
   _bitrate: number;
+  _bumper: boolean;
 
   constructor(id: string, options: PKAdOptions) {
     this._id = id;
+    this._system = options.system;
     this._url = options.url;
     this._contentType = options.contentType;
     this._title = options.title;
@@ -34,6 +37,7 @@ class Ad {
     this._width = options.width || 0;
     this._height = options.height || 0;
     this._bitrate = options.bitrate || 0;
+    this._bumper = options.bumper;
   }
 
   /**
@@ -48,7 +52,16 @@ class Ad {
   /**
    * @instance
    * @memberof Ad
-   * @return {string} - Ad ID.
+   * @return {?string} - Ad system.
+   */
+  get system(): ?string {
+    return this._system;
+  }
+
+  /**
+   * @instance
+   * @memberof Ad
+   * @return {string} - Ad content type.
    */
   get contentType(): ?string {
     return this._contentType;
@@ -120,9 +133,9 @@ class Ad {
   /**
    * @instance
    * @memberof Ad
-   * @return {string} - Whether the ad is linear.
+   * @return {boolean} - Whether the ad is linear.
    */
-  get linear(): ?boolean {
+  get linear(): boolean {
     return this._linear;
   }
 
@@ -156,6 +169,15 @@ class Ad {
   /**
    * @instance
    * @memberof Ad
+   * @return {boolean} - Whether the ad is bumper.
+   */
+  get bumper(): boolean {
+    return this._bumper;
+  }
+
+  /**
+   * @instance
+   * @memberof Ad
    * @return {string} - Whether the ad is skippable or not.
    */
   get skippable(): boolean {
@@ -165,6 +187,7 @@ class Ad {
   toJSON(): Object {
     return {
       id: this.id,
+      system: this.system,
       url: this.url,
       contentType: this.contentType,
       title: this.title,
@@ -177,7 +200,8 @@ class Ad {
       skippable: this.skippable,
       width: this.width,
       height: this.height,
-      bitrate: this.bitrate
+      bitrate: this.bitrate,
+      bumper: this.bumper
     };
   }
 }