+++ src/ads/ads-controller.js
@@ -16,17 +16,18 @@ import {Ad} from './ad';
  */
 class AdsController extends FakeEventTarget implements IAdsController {
   _player: Player;
-  _adsPluginController: IAdsController;
+  _adsPluginControllers: Array<IAdsPluginController>;
   _allAdsCompleted: boolean;
   _eventManager: EventManager;
   _adBreaksLayout: Array<number>;
   _adBreak: ?AdBreak;
   _ad: ?Ad;
+  _adPlayed: boolean;
 
-  constructor(player: Player, adsPluginController: IAdsController) {
+  constructor(player: Player, adsPluginControllers: Array<IAdsPluginController>) {
     super();
     this._player = player;
-    this._adsPluginController = adsPluginController;
+    this._adsPluginControllers = adsPluginControllers;
     this._initMembers();
     this._addBindings();
   }
@@ -83,7 +84,8 @@ class AdsController extends FakeEventTarget implements IAdsController {
    * @returns {void}
    */
   skipAd(): void {
-    this._adsPluginController.skipAd();
+    const activeController = this._adsPluginControllers.find(controller => controller.active);
+    activeController && activeController.skipAd();
   }
 
   /**
@@ -94,7 +96,8 @@ class AdsController extends FakeEventTarget implements IAdsController {
    * @returns {void}
    */
   playAdNow(adTagUrl: string): void {
-    this._adsPluginController.playAdNow(adTagUrl);
+    const activeController = this._adsPluginControllers.find(controller => controller.active);
+    activeController && activeController.playAdNow(adTagUrl);
   }
 
   _initMembers(): void {
@@ -102,6 +105,7 @@ class AdsController extends FakeEventTarget implements IAdsController {
     this._adBreaksLayout = [];
     this._adBreak = null;
     this._ad = null;
+    this._adPlayed = false;
   }
 
   _addBindings(): void {
@@ -109,15 +113,16 @@ class AdsController extends FakeEventTarget implements IAdsController {
     this._eventManager.listen(this._player, AdEventType.AD_MANIFEST_LOADED, event => this._onAdManifestLoaded(event));
     this._eventManager.listen(this._player, AdEventType.AD_BREAK_START, event => this._onAdBreakStart(event));
     this._eventManager.listen(this._player, AdEventType.AD_LOADED, event => this._onAdLoaded(event));
+    this._eventManager.listen(this._player, AdEventType.AD_STARTED, event => this._onAdStarted(event));
     this._eventManager.listen(this._player, AdEventType.AD_BREAK_END, () => this._onAdBreakEnd());
-    this._eventManager.listen(this._player, AdEventType.ALL_ADS_COMPLETED, event => this._onAllAdsCompleted(event));
+    this._eventManager.listen(this._player, AdEventType.ADS_COMPLETED, event => this._onAdsCompleted(event));
     this._eventManager.listen(this._player, AdEventType.AD_ERROR, event => this._onAdError(event));
     this._eventManager.listen(this._player, CustomEventType.PLAYER_RESET, () => this._reset());
     this._eventManager.listen(this._player, Html5EventType.ENDED, () => this._onEnded());
   }
 
   _onAdManifestLoaded(event: FakeEvent): void {
-    this._adBreaksLayout = event.payload.adBreaksPosition;
+    this._adBreaksLayout = Array.from(new Set(this._adBreaksLayout.concat(event.payload.adBreaksPosition))).sort();
     this._allAdsCompleted = false;
   }
 
@@ -130,25 +135,44 @@ class AdsController extends FakeEventTarget implements IAdsController {
     this._ad = event.payload.ad;
   }
 
+  _onAdStarted(event: FakeEvent): void {
+    this._ad = event.payload.ad;
+    this._adPlayed = true;
+  }
+
   _onAdBreakEnd(): void {
     this._adBreak = null;
     this._ad = null;
   }
 
-  _onAllAdsCompleted(event: FakeEvent): void {
-    this._allAdsCompleted = true;
-    this.dispatchEvent(event);
+  _onAdsCompleted(): void {
+    if (this._adsPluginControllers.every(controller => controller.done)) {
+      this._allAdsCompleted = true;
+      this.dispatchEvent(new FakeEvent(AdEventType.ALL_ADS_COMPLETED));
+    }
   }
 
   _onAdError(event: FakeEvent): void {
-    if (event.payload.severity === Error.Severity.CRITICAL) {
+    if (event.payload.severity === Error.Severity.CRITICAL && this._adsPluginControllers.every(controller => controller.done)) {
       this._allAdsCompleted = true;
+      if (this._adPlayed) {
+        this.dispatchEvent(new FakeEvent(AdEventType.ALL_ADS_COMPLETED));
+      }
     }
   }
 
   _onEnded(): void {
-    if (!this._allAdsCompleted && !this._adBreaksLayout.includes(-1)) {
+    if (!this._adBreaksLayout.includes(-1)) {
       this._allAdsCompleted = true;
+    } else {
+      const isBumper = controller => controller.constructor.name === 'BumperAdsController';
+      const bumperCtrl = this._adsPluginControllers.find(controller => isBumper(controller));
+      const adCtrl = this._adsPluginControllers.find(controller => !isBumper(controller));
+      const bumperCompletePromise = bumperCtrl ? bumperCtrl.onPlaybackEnded() : Promise.resolve();
+      // $FlowFixMe
+      bumperCompletePromise.finally(() => {
+        adCtrl && adCtrl.onPlaybackEnded();
+      });
     }
   }